файл переименован, сделан рефакторинг функции
